Support non-admin installs.
- Move file list generation into a helper file.
- Use mustache templating for the main XML.
- Add UI (to support non-admin installs).
